+++ collectors/jiraffe/constants.py
@@ -6,6 +6,7 @@ from osidb.helpers import get_env
 JIRA_SERVER = get_env("JIRA_URL", default="https://issues.redhat.com")
 JIRA_TOKEN = get_env("JIRA_AUTH_TOKEN")
 HTTPS_PROXY = get_env("HTTPS_TASKMAN_PROXY")
+JIRA_TASKMAN_PROJECT_KEY = get_env("JIRA_TASKMAN_PROJECT_KEY", default="OSIM")
 
 # Jira datetime format string
 JIRA_DT_FMT = "%Y-%m-%d %H:%M"
@@ -17,6 +18,9 @@ JIRA_MAX_CONNECTION_AGE = get_env("JIRA_MAX_CONNECTION_AGE")
 JIRA_BZ_ID_LABEL_RE = re.compile(r"flaw:bz#(\d+)")
 
 # Switches to turn each collector on/off
+JIRA_TASK_COLLECTOR_ENABLED = get_env(
+    "JIRA_TASK_COLLECTOR_ENABLED", default="True", is_bool=True
+)
 JIRA_TRACKER_COLLECTOR_ENABLED = get_env(
     "JIRA_TRACKER_COLLECTOR_ENABLED", default="True", is_bool=True
 )